LayoutTests/http/tests/security/cross-frame-access-object-prototype-expected.txt

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://localhost:8000/security/resources/cross-frame-iframe-for-object-prototype-test.html from frame with URL http://127.0.0.1:8000/security/cross-frame-access-object-prototype.html. Domains, protocols and ports must match.

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://localhost:8000/security/resources/cross-frame-iframe-for-object-prototype-test.html from frame with URL http://127.0.0.1:8000/security/cross-frame-access-object-prototype.html. Domains, protocols and ports must match.

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://localhost:8000/security/resources/cross-frame-iframe-for-object-prototype-test.html from frame with URL http://127.0.0.1:8000/security/cross-frame-access-object-prototype.html. Domains, protocols and ports must match.

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://localhost:8000/security/resources/cross-frame-iframe-for-object-prototype-test.html from frame with URL http://127.0.0.1:8000/security/cross-frame-access-object-prototype.html. Domains, protocols and ports must match.

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://localhost:8000/security/resources/cross-frame-iframe-for-object-prototype-test.html from frame with URL http://127.0.0.1:8000/security/cross-frame-access-object-prototype.html. Domains, protocols and ports must match.



----- test getting values cross-frame using Object.prototype -----

PASS: Object.prototype.__lookupGetter__.call(targetFrame, 'myGetter'); should be 'undefined' and is.
PASS: Object.prototype.__lookupSetter__.call(targetFrame, 'mySetter'); should be 'undefined' and is.
PASS: Object.prototype.propertyIsEnumerable.call(targetFrame, 'myProp'); should be 'false' and is.


--------
Frame: 'target'
--------

----- test that functions could not be injected cross-frame using Object.prototype -----

PASS: window.hasOwnProperty('injectedGetter') should be 'false' and is.
PASS: window.__lookupGetter__('injectedGetter') should be 'undefined' and is.
PASS: window.hasOwnProperty('injectedSetter') should be 'false' and is.
PASS: window.__lookupSetter__('injectedSetter') should be 'undefined' and is.
PASS: window.injectedGetter should be 'undefined' and is.
PASS: window.injectedSetter = 1 should be '1' and is.
